Hint bar responsive scaling and bottom-row layout polish
- Extract useViewportFitScale hook (shared by Toolbar + HintBar)
- HintBar now scales to fit the viewport like the toolbar, sits at
  bottom: 20 with height 32, and reserves horizontal margin so it
  stays clear of the "?" help button
- Move help button from (16, 16) to (20, 20) so it lines up with the
  hint bar's bottom margin
