+++ piper_draw/gui/src/components/Toolbar.tsx
@@ -8,49 +8,12 @@ import { triggerDaeImport } from "../utils/daeImport";
 import { fetchTemplateManifest, loadTemplateBlocks, type TemplateEntry } from "../utils/templates";
 import { usePreviewImages } from "./PreviewRenderer";
 import { FpsDisplay } from "./FpsCounter";
+import { useViewportFitScale } from "../hooks/useViewportFitScale";
 import type { ViewMode } from "../types";
 
-// ---------------------------------------------------------------------------
-// Responsive sizing
-// ---------------------------------------------------------------------------
-
-// Horizontal margin (px) kept between the toolbar and the viewport edges
-// when the toolbar is scaled down to fit a narrow window.
-const TOOLBAR_VIEWPORT_MARGIN_PX = 20;
-
-// Returns a CSS scale factor that keeps the toolbar at its natural size when
-// it fits in the viewport, and shrinks it just enough to fit when it doesn't.
-function useToolbarScale(toolbarRef: React.RefObject<HTMLDivElement | null>): number {
-  const [scale, setScale] = useState(1);
-  useEffect(() => {
-    let frame = 0;
-    const recompute = () => {
-      if (frame) return;
-      frame = requestAnimationFrame(() => {
-        frame = 0;
-        const node = toolbarRef.current;
-        if (!node) return;
-        // offsetWidth is the layout (untransformed) width — independent of
-        // the scale we apply, so this measurement is stable across renders.
-        const natural = node.offsetWidth;
-        if (natural === 0) return;
-        const available = window.innerWidth - TOOLBAR_VIEWPORT_MARGIN_PX;
-        setScale(Math.min(1, available / natural));
-      });
-    };
-    const node = toolbarRef.current;
-    const ro = node ? new ResizeObserver(recompute) : null;
-    if (node && ro) ro.observe(node);
-    window.addEventListener("resize", recompute);
-    recompute();
-    return () => {
-      ro?.disconnect();
-      window.removeEventListener("resize", recompute);
-      if (frame) cancelAnimationFrame(frame);
-    };
-  }, [toolbarRef]);
-  return scale;
-}
+// Horizontal margin (px) kept between fixed overlays (toolbar, hint bar) and
+// the viewport edges when they scale down to fit a narrow window.
+const VIEWPORT_FIT_MARGIN_PX = 20;
 
 // ---------------------------------------------------------------------------
 // Styles
@@ -101,7 +64,7 @@ const blockBtnStyle = (active: boolean, disabled?: boolean) => ({
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function Toolbar({ onResetCamera, controlsRef, toolbarRef, fpsRef }: { onResetCamera: () => void; controlsRef: React.RefObject<any>; toolbarRef: React.RefObject<HTMLDivElement | null>; fpsRef: React.RefObject<HTMLSpanElement | null> }) {
-  const scale = useToolbarScale(toolbarRef);
+  const scale = useViewportFitScale(toolbarRef, VIEWPORT_FIT_MARGIN_PX);
   const mode = useBlockStore((s) => s.mode);
   const setMode = useBlockStore((s) => s.setMode);
   const cubeType = useBlockStore((s) => s.cubeType);
